Basic Rust implementation of AST Nodes.
The original TypeScript implementation of this has to deal with the fact
that TS requires you to build your own tagged unions, but it also has
anonymous unions. This first-pass implementation is *mostly* just a
direct translation of those semantics to Rust. However, it also makes an
important change, because Rust `struct` types, unlike TS `interface`
types, cannot extend from other types.

As such, the `Node` type that is core to the module is "inverted":
instead of the `BaseNode` type which all other types extend from, there
is a `Node` type which has a `node` property, of which the equivalent of
the `Nodes` union type from the original is the value. This strategy
seems like the most promising approach to representing the basic
semantics of the original in a more Rustic fashion.